added game zip

+++ game/src/EnergySpark.cpp
@@ -45,7 +45,7 @@ bool EnergySpark::initialize(Object * object,  glm::vec3 target_position, unsign
 
 bool EnergySpark::update(float delta)
 {
-  float step = delta * SPEED;
+  float step = (float)delta * SPEED;
   glm::vec3 d = glm::normalize( m_target_position - m_object->getPosition() );
 
   if (glm::distance(m_target_position, m_object->getPosition()) <= step)
